+++ tspsolver.h
@@ -12,7 +12,7 @@ class TSPSolver : public QObject
     Q_OBJECT
 public:
 
-    TSPSolver(QVector<QVector<int> >& adjastancyMatrix, int nrOfCities,
+    TSPSolver(QVector<QVector<float> >& adjastancyMatrix, int nrOfCities,
               float initTemperature=0.5, float temperatureStep=0.1, float minTemperature=0.01, float mInsideLoop=1.0);
     ~TSPSolver();
     void generateStartingRoute();
@@ -23,7 +23,7 @@ public:
 
 signals:
     void newRouteComputed(QVector<QVector<int> >);
-    void finalRouteComputed(int);
+    void finalRouteComputed(float);
     void generatingRouteError();
 
 private:
@@ -31,8 +31,8 @@ private:
     int findNearestNeighbour(const int& vertexNr);
     int findNearestUnexploredVertex(const int& vertexNr);
     void findWayBackToStart(int startOfTheRoute, int lastVisitedNode);
-    int routeLength(QVector<QVector<int> > &adjacencyMatrix, QVector< QVector<int> > &route);
-    QVector<QVector<int> > mAdjacencyMatrix;
+    float routeLength(QVector<QVector<float> > &adjacencyMatrix, QVector< QVector<int> > &route);
+    QVector<QVector<float> > mAdjacencyMatrix;
     int mNrOfCities;
     float mInitTemperature;
     float mTemperatureStep;
